refactor: rename, relocate, reimport modules (added engine folder)
