--- nyx/ecs/component/components.py
@@ -1,44 +1,0 @@
-from abc import ABC
-from typing import Type
-import numpy as np
-
-
-class NyxComponent(ABC):
-    """Abstract base class of all components in NyxEngine."""
-
-
-class TransformComponent(NyxComponent):
-    def __init__(self, x: int, y: int):
-        self.x = x
-        self.y = y
-
-
-class GraphicComponent(NyxComponent):
-    def __init__(self, graphic: np.ndarray):
-        if not isinstance(graphic, np.ndarray) or graphic.dtype != np.uint8:
-            raise ValueError("Graphic must be a numpy ndarray of dtype 'uint8'")
-        self.graphic = graphic
-
-
-class TileSetComponent(NyxComponent):
-    def __init__(self, tile_dimension: int = 16):
-        self.tiles = {}
-        self.tile_dimension = (
-            tile_dimension
-            if isinstance(tile_dimension, int) and tile_dimension > 0
-            else 16
-        )
-
-
-class TileMapComponent(NyxComponent):
-    def __init__(self, tilemap: np.ndarray, tile_width: int = 1, tile_height: int = 1):
-        if not isinstance(tilemap, np.ndarray) or tilemap.dtype != np.uint8:
-            raise ValueError("Tilemap must be a NumPy `ndarray` of `dtype` 'uint8'")
-
-        self.tilemap = tilemap
-        self.tile_width = tile_width
-        self.tile_height - tile_height
-
-    def __str__(self):
-        """Return the class name (which is same as component name) when called as string."""
-        return f"{Type[self]}"
