--- nyx/ecs/system/nyx_system_base.py
@@ -1,7 +1,0 @@
-from abc import ABC
-from nyx.ecs.nyx_entity_manager import NyxEntityManager
-
-
-class NyxSystem(ABC):
-    def __init__(self, manager: NyxEntityManager):
-        self.manager = manager
