+++ nyx/engine/ecs/system/nyx_system_base.py
@@ -1,0 +1,7 @@
+from abc import ABC
+from nyx.engine.ecs.nyx_entity_manager import NyxEntityManager
+
+
+class NyxSystem(ABC):
+    def __init__(self, ecs: NyxEntityManager):
+        self.ecs = ecs
